Update README.md
After cleaning up the files, updating some of the code, and making some generalized tweaks and quality of life improvements to the organization of my files inside the repo, the instructions needed updating. Everything *should* be more accurate now and reflect the current state of my repo, as well as have the correct instructions for using my files! :3