Create alacritty.toml
alacritty has moved to .toml files for configuration rather than .yml, i migrated the original .yml file and fixed outdated functions that were called by the configuration. I will update the colors and rework the file a bit in the future, but for now its the same file as before with a few things removed (footer/ color scheme calls are not supported by the toml format).