Update  config
A massive update to my i3 config file. I have been away from i3 and arch for quite some time, but i have returned and fixed some of the issues i had prior, as well as changed the monitor layout since i have a 3rd, vertical monitor now. I need to clean up the file a little bit, this is the same file that im running, but im not completely happy with it and plan to update it soon(ish, maybe?). That or ill forget about it for another year. Who knows?